Moving group aggregate operators to separate file #149
